# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Tutor Marketplace CRM & Lead Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tutors:
    get:
      operationId: listTutors
      tags: [tutors]
      summary: List tutors
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: subject
          in: query
          schema:
            type: string
        - name: area
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of tutors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tutor"
    post:
      operationId: createTutor
      tags: [tutors]
      summary: Register a tutor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TutorInput"
      responses:
        "201":
          description: Tutor created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tutor"
  /tutors/{id}:
    get:
      operationId: getTutor
      tags: [tutors]
      summary: Get tutor profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tutor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TutorDetail"
        "404":
          description: Not found
    patch:
      operationId: updateTutor
      tags: [tutors]
      summary: Update tutor profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TutorUpdate"
      responses:
        "200":
          description: Updated tutor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tutor"

  # Requirements
...
components:
  schemas:
...
    Tutor:
      type: object
      required: [id, name, phone, qualifications, subjects, areas, status, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        qualifications:
          type: string
        experience:
          type: ["integer", "null"]
        subjects:
          type: array
          items:
            type: string
        areas:
          type: array
          items:
            type: string
        grades:
          type: array
          items:
            type: string
        expectedFee:
          type: ["number", "null"]
        status:
          type: string
          enum: [active, inactive, blacklisted]
        gender:
          type: ["string", "null"]
        teachingMode:
          type: string
          enum: [home, online, both]
        createdAt:
          type: string
    TutorDetail:
      allOf:
        - $ref: "#/components/schemas/Tutor"
        - type: object
          properties:
            applications:
              type: array
              items:
                $ref: "#/components/schemas/Application"
            demos:
              type: array
              items:
                $ref: "#/components/schemas/Demo"
    TutorInput:
      type: object
      required: [name, phone, qualifications, subjects, areas, teachingMode]
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        qualifications:
          type: string
        experience:
          type: integer
        subjects:
          type: array
          items:
            type: string
        areas:
          type: array
          items:
            type: string
        grades:
          type: array
          items:
            type: string
        expectedFee:
          type: number
        gender:
          type: string
        teachingMode:
          type: string
          enum: [home, online, both]
        bio:
          type: string
    TutorUpdate:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        qualifications:
          type: string
        experience:
          type: integer
        subjects:
          type: array
          items:
            type: string
        areas:
          type: array
          items:
            type: string
        grades:
          type: array
          items:
            type: string
        expectedFee:
          type: number
        gender:
          type: string
        teachingMode:
          type: string
          enum: [home, online, both]
        status:
          type: string
          enum: [active, inactive, blacklisted]
        bio:
          type: string
...
    Application:
      type: object
      required: [id, requirementId, tutorId, status, createdAt]
      properties:
        id:
          type: integer
        requirementId:
          type: integer
        tutorId:
          type: integer
        tutorName:
          type: string
        tutorPhone:
          type: string
        status:
          type: string
          enum: [pending, shortlisted, rejected, selected]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
...
    Demo:
      type: object
      required: [id, requirementId, tutorId, scheduledAt, status, createdAt]
      properties:
        id:
          type: integer
        requirementId:
          type: integer
        tutorId:
          type: integer
        tutorName:
          type: string
        parentId:
          type: integer
        parentName:
          type: string
        parentPhone:
          type: string
        subject:
          type: string
        area:
          type: string
        scheduledAt:
          type: string
        status:
          type: string
          enum: [scheduled, completed, cancelled, rescheduled]
        feedback:
          type: ["string", "null"]
        outcome:
          type: ["string", "null"]
        assignedTo:
          type: ["integer", "null"]
        createdAt:
          type: string
...