# lib/api-zod/src/generated/types/tutorUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Tutor Marketplace CRM & Lead Management API
 * OpenAPI spec version: 0.1.0
 */
import type { TutorUpdateStatus } from './tutorUpdateStatus';
import type { TutorUpdateTeachingMode } from './tutorUpdateTeachingMode';

export interface TutorUpdate {
  name?: string;
  phone?: string;
  email?: string;
  qualifications?: string;
  experience?: number;
  subjects?: string[];
  areas?: string[];
  grades?: string[];
  expectedFee?: number;
  gender?: string;
  teachingMode?: TutorUpdateTeachingMode;
  status?: TutorUpdateStatus;
  bio?: string;
}